refactor(web): finish the PageShell migration cleanup
Rename office-topbar.tsx to office-shell.tsx so the file matches the
OfficeShell it exports, and update both paint paths.

Drop the last two nested <main> elements (tasks list, office run detail).
AppShell owns that landmark, and the migrated pages already say so.

Rename Linear's local page frame to LinearFrame so it stops shadowing the
shared PageShell it wraps.

Give the icon-only topbar home crumb a translated accessible name from
sidebar:home — the same catalog key the nav manifest's Home row resolves —
instead of a hardcoded "Home".

Trim the MOBILE_MENU_EXEMPTIONS doc block now that the list is empty, and
name its guardrail test for what it actually checks.
