+++ apps/web/app/office/components/office-shell.tsx
