+++ apps/web/lib/navigation/core-destinations.test.ts
@@ -94,7 +94,11 @@ describe("navigation coverage guardrails", () => {
     expect(ids(primary)).toEqual(["home", "tasks"]);
   });
 
-  it("keeps the exemption list honest", () => {
+  // Shape guard for whatever comes back into MOBILE_MENU_EXEMPTIONS: the list is
+  // empty today (the coverage test above is what carries the weight), so this
+  // asserts nothing until an entry is added — at which point it must name a real
+  // destination that genuinely has no mobileMenu surface, and say why.
+  it("holds any future exemption to naming a real destination and a reason", () => {
     for (const [id, reason] of Object.entries(MOBILE_MENU_EXEMPTIONS)) {
       const destination = APP_DESTINATIONS.find((entry) => entry.id === id);
       expect(destination, `${id} is exempt but is not a destination`).toBeTruthy();
