+++ apps/web/lib/navigation/surface-policy.ts
@@ -52,15 +52,13 @@ export const MOBILE_MENU_UTILITY_SECTIONS: StaticNavSection[] = ["insights", "ut
 
 /**
  * Destinations deliberately not offered in the mobile menu, each with the mobile
- * affordance that owns it instead. Everything else must be offered there: the
- * sidebar is hidden below `md`, so an omission means unreachable on a phone —
- * which is exactly how `/stats` ended up with no phone entry point.
- * `core-destinations.test.ts` fails on any omission that is not listed here.
- *
- * Currently empty: every destination is offered. A surface that already owns a
- * given affordance opts out at its own call site instead (the kanban drawer
- * renders `AppNavSections` with `omitSections={["primary"]}` because its brand
- * link and View toggle cover Home and Tasks).
+ * affordance that owns it instead — the sidebar is hidden below `md`, so an
+ * omission means unreachable on a phone, which is exactly how `/stats` ended up
+ * with no phone entry point. `core-destinations.test.ts` fails on any omission
+ * that is not listed here. Empty today: every destination is offered, and a
+ * surface that already covers one opts out at its own call site instead (the
+ * kanban drawer passes `omitSections={["primary"]}` because its brand link and
+ * View toggle cover Home and Tasks).
  */
 export const MOBILE_MENU_EXEMPTIONS: Record<string, string> = {};
